backend/requirements.txt
# Kavita SafeUploader Backend Dependencies
# Python 3.12+ (Ubuntu 24.04 LTS)

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Configuration & Environment
pyyaml==6.0.2
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Security & Validation
python-magic==0.4.27
slowapi==0.1.9
pydantic==2.9.2
pydantic-settings==2.5.2

# File Operations & Hashing
aiofiles==24.1.0

# HTTP Client (for VirusTotal API - Step 2)
httpx==0.27.2

# Step 3: Metadata extraction and preview generation
ebooklib==0.18
PyMuPDF==1.24.14  # PDF preview and metadata
Pillow==11.0.0  # Image processing
rarfile==4.2  # CBR extraction
pdfminer.six==20231228  # PDF text extraction (fallback)

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2  # For test client

